+++ lib/tr_extension.dart
@@ -9,11 +9,9 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 export 'package:tr_extension/tr_extension.dart';
 
-part 'src/class.dart';
 part 'src/delegate.dart';
 part 'src/exception.dart';
 part 'src/extension.dart';
-part 'src/install.dart';
 
 typedef Translations = Map<String, Map<String, String>>;
 typedef TranslationsArgs = Map<String, Map<String, ArgReplacer>>;