+++ test/tr_extension_test.dart
@@ -3,8 +3,8 @@ import 'dart:ui';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:tr_extension/tr_extension.dart';
 
-
 void main() {
+  final tr = TrDelegate();
   group('strings', () {
     test('Fallback Pattern', () {
       const words = '1.2.3.4.5';
@@ -15,7 +15,7 @@ void main() {
       expect(subwords.last, '1');
     });
 
-    String tr() {
+    String translate() {
       const words = '1.2.3.4.5';
       final subwords = words.subWords('.');
 
@@ -27,7 +27,7 @@ void main() {
     }
 
     test('tr', () {
-      expect(tr(), '1.2.3.4.5');
+      expect(translate(), '1.2.3.4.5');
     });
 
     test('Dynamic Token Replacement', () {
@@ -41,27 +41,29 @@ void main() {
         'user.items': 'You have items!',
       };
 
-      Tr.to.putTranslations(const Locale('en'), keys);
+      const locale = Locale('en');
+      tr.setTranslations(locale, keys);
+      tr.setLocale(locale);
 
-      final items = Tr.to.translate('user.items.0');
+      final items = tr.translate('user.items.0');
       expect(items, 'You have zero items :(');
 
-      final item = Tr.to.translate('user.items.1');
+      final item = tr.translate('user.items.1');
       expect(item, 'You have exactly one item!');
 
-      final item2 = Tr.to.translate('user.items.2');
+      final item2 = tr.translate('user.items.2');
       expect(item2, 'You have 2 items!');
 
-      final item3 = Tr.to.translate('user.items.Art.3');
+      final item3 = tr.translate('user.items.Art.3');
       expect(item3, 'Art have 3 items!');
 
-      final item4 = Tr.to.translate('user.items.now.do.we');
+      final item4 = tr.translate('user.items.now.do.we');
       expect(item4, 'we may do this now');
 
-      final item5 = Tr.to.translate('user.items.look.bet.dancefloor.on');
+      final item5 = tr.translate('user.items.look.bet.dancefloor.on');
       expect(item5, 'I bet you look good on the dancefloor');
 
-      final item6 = Tr.to.translate('user.items');
+      final item6 = tr.translate('user.items');
       expect(item6, 'You have items!');
     });
   });